refactor: move site-config.ts into lib/content directory
