+++ components/home/homepage-welcome/index.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useRef, useState } from "react";
 
+import siteConfig from "~/lib/content/site-config";
 import useIntersectionObserver from "~/lib/hooks/use-intersection";
-import siteConfig from "~/lib/site-config";
 import { attributes, html } from "~/lib/welcome-message";
 
 import { BasicDialog } from "./dialog";
