+++ components/home/homepage-whatsapp-cta.tsx
@@ -1,4 +1,4 @@
-import siteConfig from "~/lib/site-config";
+import siteConfig from "~/lib/content/site-config";
 
 import { PrimaryAnchorButton } from "../ui/button";
 import { WhatsAppIcon } from "../ui/icons";
