+++ pages/isolasi-mandiri.tsx
@@ -3,8 +3,8 @@ import { Page } from "~/components/layout/page";
 import { PageContent } from "~/components/layout/page-content";
 import { PageHeader } from "~/components/layout/page-header";
 import StackedLink from "~/components/layout/stacked-link";
+import siteConfig from "~/lib/content/site-config";
 import isoman, { Category } from "~/lib/isoman-contents";
-import siteConfig from "~/lib/site-config";
 
 import { GetStaticProps } from "next";
 import { NextSeo } from "next-seo";
